chore: move functions from commoon_types.h/c to conversions.c/h
