+++ lte/gateway/c/core/oai/common/conversions.c
@@ -38,6 +38,7 @@
 #include <ctype.h>
 
 #include "conversions.h"
+#include "log.h"
 
 static const char hex_to_ascii_table[16] = {
     '0', '1', '2', '3', '4', '5', '6', '7',
@@ -196,4 +197,161 @@ imsi64_t amf_imsi_to_imsi64(const imsi_t* const imsi) {
   }
 
   return imsi64;
+}
+
+/* Clear GUTI without free it */
+void clear_guti(guti_t* const guti) {
+  memset(guti, 0, sizeof(guti_t));
+  guti->m_tmsi = INVALID_TMSI;
+}
+/* Clear IMSI without free it */
+void clear_imsi(imsi_t* const imsi) {
+  memset(imsi, 0, sizeof(imsi_t));
+}
+/* Clear IMEI without free it */
+void clear_imei(imei_t* const imei) {
+  memset(imei, 0, sizeof(imei_t));
+}
+/* Clear IMEISV without free it */
+void clear_imeisv(imeisv_t* const imeisv) {
+  memset(imeisv, 0, sizeof(imeisv_t));
+}
+
+//------------------------------------------------------------------------------
+bstring fteid_ip_address_to_bstring(const struct fteid_s* const fteid) {
+  bstring bstr = NULL;
+  if (fteid->ipv4) {
+    bstr = blk2bstr(&fteid->ipv4_address.s_addr, 4);
+  } else if (fteid->ipv6) {
+    bstr = blk2bstr(&fteid->ipv6_address, 16);
+  } else {
+    char avoid_seg_fault[4] = {0, 0, 0, 0};
+    bstr                    = blk2bstr(avoid_seg_fault, 4);
+  }
+  return bstr;
+}
+
+void get_fteid_ip_address(
+    const struct fteid_s* const fteid, ip_address_t* const ip_address) {
+  if (fteid->ipv4) {
+    ip_address->pdn_type = IPv4;
+    memcpy(&ip_address->address.ipv4_address, &fteid->ipv4_address, 4);
+  }
+
+  if (fteid->ipv6) {
+    ip_address->pdn_type = IPv6;
+    memcpy(&ip_address->address.ipv6_address, &fteid->ipv6_address, 16);
+  }
+
+  if (fteid->ipv4 && fteid->ipv6) {
+    ip_address->pdn_type = IPv4_AND_v6;
+  }
+}
+
+//------------------------------------------------------------------------------
+bstring ip_address_to_bstring(const ip_address_t* ip_address) {
+  bstring bstr = NULL;
+  switch (ip_address->pdn_type) {
+    case IPv4:
+      bstr = blk2bstr(&ip_address->address.ipv4_address.s_addr, 4);
+      break;
+    case IPv6:
+      bstr = blk2bstr(&ip_address->address.ipv6_address, 16);
+      break;
+    case IPv4_AND_v6:
+      bstr = blk2bstr(&ip_address->address.ipv4_address.s_addr, 4);
+      bcatblk(bstr, &ip_address->address.ipv6_address, 16);
+      break;
+    case IPv4_OR_v6:
+      // do it like that now, TODO
+      bstr = blk2bstr(&ip_address->address.ipv4_address.s_addr, 4);
+      break;
+    default: {}
+  }
+  return bstr;
+}
+
+//------------------------------------------------------------------------------
+void bstring_to_ip_address(bstring const bstr, ip_address_t* const ip_address) {
+  if (bstr) {
+    switch (blength(bstr)) {
+      case 4:
+        ip_address->pdn_type = IPv4;
+        memcpy(&ip_address->address.ipv4_address, bstr->data, blength(bstr));
+        break;
+      case 16:
+        ip_address->pdn_type = IPv6;
+        memcpy(&ip_address->address.ipv6_address, bstr->data, blength(bstr));
+        break;
+        break;
+      case 20:
+        ip_address->pdn_type = IPv4_AND_v6;
+        memcpy(&ip_address->address.ipv4_address, bstr->data, 4);
+        memcpy(&ip_address->address.ipv6_address, &bstr->data[4], 16);
+        break;
+      default: {}
+    }
+  }
+}
+
+//------------------------------------------------------------------------------
+void copy_paa(paa_t* paa_dst, paa_t* paa_src) {
+  memcpy(paa_dst, paa_src, sizeof(paa_t));
+}
+
+//------------------------------------------------------------------------------
+bstring paa_to_bstring(const paa_t* paa) {
+  bstring bstr = NULL;
+  switch (paa->pdn_type) {
+    case IPv4:
+      bstr = blk2bstr(&paa->ipv4_address.s_addr, 4);
+      break;
+    case IPv6:
+      if (paa->ipv6_prefix_length == IPV6_PREFIX_LEN) {
+        bstr = blk2bstr(&paa->ipv6_address, paa->ipv6_prefix_length / 8);
+      } else {
+        OAILOG_ERROR(
+            LOG_COMMON, "Invalid ipv6_prefix_length : %u\n",
+            paa->ipv6_prefix_length);
+      }
+      break;
+    case IPv4_AND_v6:
+      if (paa->ipv6_prefix_length == IPV6_PREFIX_LEN) {
+        bstr = blk2bstr(&paa->ipv6_address, paa->ipv6_prefix_length / 8);
+        bcatblk(bstr, &paa->ipv4_address, 4);
+      } else {
+        OAILOG_ERROR(
+            LOG_COMMON, "Invalid ipv6_prefix_length : %u\n",
+            paa->ipv6_prefix_length);
+      }
+      break;
+    case IPv4_OR_v6:
+      // do it like that now, TODO
+      bstr = blk2bstr(&paa->ipv4_address.s_addr, 4);
+      break;
+    default: {}
+  }
+  return bstr;
+}
+
+void bstring_to_paa(const bstring bstr, paa_t* paa) {
+  if (bstr) {
+    switch (blength(bstr)) {
+      case 4:
+        paa->pdn_type = IPv4;
+        memcpy(&paa->ipv4_address, bstr->data, blength(bstr));
+        break;
+      case 8:
+        paa->pdn_type = IPv6;
+        memcpy(&paa->ipv6_address, bstr->data, blength(bstr));
+        break;
+      case 12:
+        paa->pdn_type = IPv4_AND_v6;
+        memcpy(&paa->ipv4_address, bstr->data, 4);
+        memcpy(&paa->ipv6_address, &bstr->data[4], 8);
+        break;
+      default:
+        break;
+    }
+  }
 }
