+++ lte/gateway/c/core/oai/tasks/sgw/spgw_state_converter.cpp
@@ -18,6 +18,7 @@
 #include "spgw_state_converter.h"
 
 extern "C" {
+#include "conversions.h"
 #include "dynamic_memory_check.h"
 #include "sgw_context_manager.h"
 }
